refactor: move indexing utils to dedicated folder
